// src/metadata/enums.ts
export enum Status {
    ACTIVE = 'ACTIVE',
    INACTIVE = 'INACTIVE'
}

export enum Severity {
    MILD = 'MILD',
    MODERATE = 'MODERATE',
    SEVERE = 'SEVERE'
}

export enum Speciality {
    GENERAL = 'GENERAL',
    SURGEON = 'SURGEON',
    PEDIATRICIAN = 'PEDIATRICIAN',
    CARDIOLOGIST = 'CARDIOLOGIST'
}